requirements.txt
Django==1.11.1
PyMySQL==0.7.11
Pillow
djangorestframework
django-filter
django-rest-knox